+++ api/V0ScaleZookeeper.go
@@ -6,6 +6,8 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/sirupsen/logrus"
+
+	mesos "../mesos"
 )
 
 // V0ScaleZookeeper will scale the zookeeper service
@@ -18,11 +20,38 @@ func V0ScaleZookeeper(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	d := []byte("{\"nok\"}")
+	d := []byte("nok")
 
 	if vars["count"] != "" {
-		config.ZookeeperMax, _ = strconv.Atoi(vars["count"])
-		d = []byte("{\"ok\"}")
+		newCount, _ := strconv.Atoi(vars["count"])
+		oldCount := config.ZookeeperMax
+		i := (newCount - oldCount) * -1
+		// Scale Up
+		if newCount > oldCount {
+
+			config.ZookeeperMax = newCount
+
+			logrus.Info("Zookeeper Scale Up ", i)
+			for x := oldCount; x < newCount; x++ {
+				mesos.GetZookeeperServerString(x)
+				mesos.StartZookeeper(x)
+			}
+		}
+
+		// Scale Down
+		if newCount < oldCount {
+			logrus.Info("Zookeeper Scale Down ", i)
+
+			for x := newCount; x < oldCount; x++ {
+				task := mesos.StatusZookeeper(x)
+				id := *task.Status.TaskId.Value
+				mesos.Kill(id)
+			}
+		}
+
+		config.ZookeeperMax = newCount
+
+		d = []byte(strconv.Itoa(config.ZookeeperMax))
 	}
 
 	logrus.Debug("HTTP GET V0ScaleZookeeper: ", string(d))